Create TypeScript index.d.ts definitions
Adds typescript support for the component which embeds Elm code in
React. Also provides helper types for extracting the Ports and Flags
used by an Elm program.

These type declarations are useful in contexts where the TypeScript
type declaration of the underlying Elm program is either hand-written,
or generated by a tool such as `elm-typescript-interop`.